backend/pyproject.toml
[project]
name = "direphish-backend"
version = "0.1.0"
description = "DirePhish — Swarm-predict what goes wrong next"
requires-python = ">=3.11"
license = { text = "AGPL-3.0" }
authors = [
    { name = "raxIT Labs" }
]

dependencies = [
    "flask>=3.0.0",
    "gunicorn>=23.0.0",
    "flask-cors>=6.0.0",
    "openai>=1.0.0",
    "zep-cloud==3.13.0",
    "graphiti-core[google-genai]==0.28.2",
    "kuzu>=0.11.0",
    "camel-oasis==0.2.5",
    "camel-ai==0.2.78",
    "PyMuPDF>=1.24.0",
    "charset-normalizer>=3.0.0",
    "chardet>=5.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "crucible-sim @ git+https://github.com/raxITlabs/crucible.git",
    "opentelemetry-api>=1.40.0",
    "opentelemetry-sdk>=1.40.0",
    "opentelemetry-exporter-otlp-proto-http>=1.40.0",
    # Firestore Vector Search (replaces Graphiti + Zep for simulation memory)
    "google-cloud-firestore>=2.16.0",
    "google-genai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pipreqs>=0.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.uv]
# graphiti-core requires neo4j>=5.26.0; override camel-oasis's strict neo4j==5.23.0 pin
override-dependencies = [
    "neo4j>=5.26.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.hatch.build.targets.wheel]
packages = ["app"]
